Overwriting bonus points on automated calculation
Once bonuspoints are calculated automatically,
the resulting value is saved in the appropriate field of team object.
This feature is needed in case if you have to fallback to manual bonus points
after several fight ended.
With this feature, you will not have to manually recalculate all the bonus points
for all the past fights (but feel free to do it if you want).

Note that swutching bonus points back to auto rewrites all your manual edits.