[refactoring] Turn the list of tournaments into a separate setting - see #78
The list of tournaments used to be appended to the INSTALLED_APPS tuple.
Now, the list is a separate tuple in the settings.
The corresponding URLs are supported automatically.

You don't have to edit `urls.py` anymore:
all the INSTALLED_TOURNAMENTS will get their URLs automatically,
and the first of them will be the default one
(i.e. it will be displayed on the main page)