[#235] Sort the problems in old-fashioned lists
This commit breaks HTML equivalence (up to the order of items in `<ul>`s).

Before the #235-related-refactoring has started,
there were two completely different functions for listing forbidden problems
(see also #133):
a method of a Round object (in  `models.py`)
and a standalone function  (in `tactics.py`).

The first one, which is edited in the commit,
outputs the problems without any visible sane order.
So, it would be very difficult to compare the results
of the old and the new functions.
As a workaround, we sort the problems in the output of the old function
to be sorted by number inside each reason.