[fix #206] Update the scores of the involved Teams when a Round is deleted
Initially, deleting a Round did not lead to the update,
because deleting a Round is a rare action.
However, during TST-2020 the organizer tried to add a test Round
before the tournament had started and then deleted the Round.
The ranking wasn't set to zeros, and `update_all` did not help.
Thus, the fix was crucial.