[fix] Don't crash when tactical rejections are disaled
We used to create database fields for some SiteConfiguration options
only if these options make sense.
However, there turned out to be some pieces of code
that try to read such values, that obviously lead to an error.

Thus, we decided to create the fileds anyway,
but just make them inaccessible if they don't make sense.